Floyd Warshall complete

+++ src/uniandes/algorithms/minimumcostpath/BellmanFordsAlgorithm.java
@@ -2,7 +2,7 @@ package uniandes.algorithms.minimumcostpath;
 
 public class BellmanFordsAlgorithm implements MinimumCostPathAlgorithm {
 
-	public int[][] findPaths(int[][] input) {
+	public Double[][] findPaths(Double[][] input) {
 		return null;
 	}
 }
